+++ gradle.properties
@@ -8,9 +8,9 @@ org.gradle.jvmargs=-Xmx1G
 	loader_version=0.13.3
 
 # Mod Properties
-	mod_version = 1.0.0
-	maven_group = com.example
-	archives_base_name = fabric-example-mod
+	mod_version = 0.0.1
+	maven_group = net.tentaken
+	archives_base_name = tutorialmod
 
 # Dependencies
 	fabric_version=0.47.8+1.18.2
