+++ src/main/java/net/tentaken/tutorialmod/TutorialMod.java
@@ -1,10 +1,10 @@
-package net.fabricmc.example;
+package net.tentaken.tutorialmod;
 
 import net.fabricmc.api.ModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ExampleMod implements ModInitializer {
+public class TutorialMod implements ModInitializer {
 	// This logger is used to write text to the console and the log file.
 	// It is considered best practice to use your mod id as the logger's name.
 	// That way, it's clear which mod wrote info, warnings, and errors.
