+++ src/main/resources/tutorialmod.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "net.fabricmc.example.mixin",
+  "package": "net.tentaken.tutorialmod.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
   ],
